Rename include header for native
