+++ plugin-dev/Source/Sentry/Private/GenericPlatform/Convenience/GenericPlatformSentryInclude.h
