+++ plugin-dev/Source/Sentry/Private/GenericPlatform/GenericPlatformSentryBreadcrumb.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "Convenience/SentryInclude.h"
+#include "Convenience/GenericPlatformSentryInclude.h"
 
 #include "Interface/SentryBreadcrumbInterface.h"
